openshorts container renamed
