+++ apps/hyperdrive-trading/src/ui/portfolio/Portfolio.tsx
@@ -3,7 +3,7 @@ import { ReactElement } from "react";
 import { TabsTwo } from "src/ui/base/components/Tabs/TabsTwo";
 import { OpenLongsContainer } from "src/ui/hyperdrive/longs/OpenLongsTable/OpenLongsTableDesktop";
 import { LpAndWithdrawalSharesContainer } from "src/ui/hyperdrive/lp/LpAndWithdrawalSharesTable/LpAndWithdrawalSharesTable";
-import { OpenShortsContainer } from "src/ui/hyperdrive/shorts/OpenShortsTable/OpenShortsTableDesktopTwo";
+import { OpenShortsContainer } from "src/ui/hyperdrive/shorts/OpenShortsTable/OpenShortsTableDesktop";
 import { PORTFOLIO_ROUTE } from "src/ui/portfolio/routes";
 
 export function Portfolio(): ReactElement {
